Basic Object

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -7,19 +7,19 @@ package bgu.spl.mics.application.objects;
 public class CloudPoint {
 
 // --------------------- fields -------------------------
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
 // --------------------- constructor --------------------
-    public CloudPoint(int x, int y){
+    public CloudPoint(double x, double y){
         this.x = x;
         this.y = y;
     }
 
 // --------------------- methods ------------------------
 //Getters
-    public int getX() {return x;}
-    public int getY() {return y;}
+    public double getX() {return x;}
+    public double getY() {return y;}
 
 //OtherMethods
     public void update(CloudPoint newPoint) {
